Add processing of 'identity' parameter for mcollective config
Nailgun-agent provided the parameter for the config and restarts
mcollective after update. But in some cases (see description of the
related bug) such behavior may cause deployment failure. So now the data
is supplied by astute in provision info and is placed into config on its
creation as other parameters.

Change-Id: I3670e571c13808da2b54bd6238d228e7cdb0ef96
Related-Bug: #1518306
Depends-On: I753eb76ed9c3b80f249c0c4b86ef48ef49274990