Put py34 first in the env order of tox, and remove py26/py33
To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then,
run py27. This patch puts py34 first on the tox.ini list of envs to
avoid this problem to happen.

This patch also removes the support for py26/py33 as they would not be
supported from Mitaka Release.


Change-Id: I248fd2e4e28afb605528e4ec060e28664bc82b5a
Closes-bug: #1489059