Use any disk for /boot regardless of its size
Apparently, the disks which're bigger than 2T were excluded from list
of bootable disks if possible. Just because in the past, those disks
were unrecognized by BIOS (or UEFI under CSM).

In fact, it was just misconfiguration of RAID controller or BIOS
itself.

GRUB uses BIOS INT13h in order to find all available disks.
Therefore, unless particular disk is not configured as 'bootable',
there's no change for GRUB to find it.

One should configure hardware in the following way assuming that
the first disk (hd0) is bootable and is used for operating system
purposes.
In case of hardware RAID, FC multipath or any other HBA, the disk
(or lun, or whatever) which was configured as 'bootable' will be
reported as hd0 via INT13h. So, GRUB will be able to boot from it.

DocImpact
Closes-Bug: #1588260

Change-Id: I7bc729ffafa3b9d6bfe8521fa38599d36d02f7e1