Fix issue with disks without partitions
When "patition_policy" set to "verify" into deployment config, we are
collecting info about existing disk and partitions to compare it with
requested partitions scheme. But command "parted" can't process disks
without partitions table. It ends with error code 1 and does not have
very detailed/correct error message.

The simplest way to fix this issue is to skip such disks. But on this
stage we do not have enouph info to make this decision. Incorrect
decision here can lead to incorrect internal representation of existing
partitions and, in worst case, to data loss.

This patch increases complexity of partition analysis - it adds one
more tool(lsblk) to grab "basic" disk info and check if partition table
is present.

PS Solution is not optimal. Partition analyse process must be rewriten.

Change-Id: Ie7fe031244f8a8f71cbd43ebaa43821af1529fb1