Avoid persistent storage modification by request
Ironic driver support "check and live as is" partitioning operation. It
is used to in "redeploy" node after hardware failure. Ususally some
files are changes since time when node was deployed. And usually we need
to keep this files "as is" too.

The problem appers in "do_install_os" deployment step. It revert all
files to state stored into deploing image.

This update disable "do_install_os" when "check and live as is"
partitioning operation is requested.

Change-Id: I978059ec28444389e5a1b18ae0a20af0e1a71a45